requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
defusedxml==0.6.0
Django==2.2.1
django-modeltranslation==0.13.1
future==0.17.1
gunicorn==19.9.0
http-parser==0.8.3
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
pbr==5.2.0
Pillow==6.0.0
git+https://github.com/probachaiyu/pipedrive-python.git
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
simplejson==3.16.0
six==1.12.0
socketpool==0.5.3
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
zdesk==2.7.1
django-rest-framework==0.1.0
